+++ drift_dev/test/writer/data_class_writer_test.dart
@@ -212,6 +212,61 @@ class MyTable extends Table {
   }''')),
     }, result.dartOutputs, result.writer);
   });
+
+  test('generates correct companions for modular row classes', () async {
+    final result = await emulateDriftBuild(
+      inputs: {
+        'a|lib/a.dart': '''
+import 'package:drift/drift.dart';
+
+@UseRowClass(Item, generateInsertable: true)
+class ItemTable extends Table {
+  IntColumn get id => integer().autoIncrement()();
+}
+
+class Item {
+  final int id;
+  Item(this.id);
+}
+''',
+      },
+      modularBuild: true,
+    );
+
+    checkOutputs(
+      {
+        'a|lib/a.drift.dart': decodedMatches(
+          allOf(
+            // The toString() definition for companions was broken and included
+            // the import prefix of the companion.
+            contains("StringBuffer('ItemTableCompanion(')"),
+
+            // The extension should also reference the row class correctly
+            contains(r'''
+class _$ItemInsertable implements i0.Insertable<i1.Item> {
+  i1.Item _object;
+  _$ItemInsertable(this._object);
+  @override
+  Map<String, i0.Expression> toColumns(bool nullToAbsent) {
+    return i2.ItemTableCompanion(
+      id: i0.Value(_object.id),
+    ).toColumns(false);
+  }
+}
+
+extension ItemToInsertable on i1.Item {
+  _$ItemInsertable toInsertable() {
+    return _$ItemInsertable(this);
+  }
+}
+'''),
+          ),
+        ),
+      },
+      result.dartOutputs,
+      result.writer,
+    );
+  });
 }
 
 class _GeneratesConstDataClasses extends Matcher {